+++ docs/changelog.rst
@@ -19,7 +19,7 @@ Currently in development.
 - Split setting for showing topics in navbar for actions and instruments tabs
 - Added webhook type for changes to object permissions
 - Added previous versions to documentation
-- Added copy buttons to regular and list-style arrays
+- Added copy and move buttons to arrays
 - Fixed array tables